Исправил тест конструктора и вынес в After удаление пользователя

+++ src/test/java/LogOutTest.java
@@ -8,6 +8,7 @@ import PageObjects.PersonalAreaPage;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -34,6 +35,12 @@ public class LogOutTest extends Annotations {
         objPersonalAreaPage.logOutButtonClick();
         boolean isVisibleLoginPage = objLoginPage.visibleLoginPage();
         Assert.assertTrue(isVisibleLoginPage);
-        userClient.deleteUser(accessToken);
+    }
+
+    @After
+    public void deleteUer() {
+        if (accessToken != null) {
+            userClient.deleteUser(accessToken);
+        }
     }
 }
